requirements/dev.txt
-r ./common.txt

# Django debug information
django-debug-toolbar==4.1.0

# Extensions for Django
django-extensions==3.2.3